audio: Implement adev_set_mic_mute
The mic mute function sets to 0 the voice uplink volume if state
parametter is true, else it sets it to default volume (0dB or 6dB).

We mute the voice uplink in the voice call mode, else
in the multimedia mode we mute all mics.

To use this AESS mute instead of the default mute made by the modem,
set to "true" the send_mic_mute_to_AudioManager resource in the
packages/apps/Phone/res/values/config.xml file.

Change-Id: I321049e02a8b96c81a7d7281666818237989905e
Signed-off-by: Cedric Poignet <c-poignet@ti.com>